GeoCoordinate double to decimal
Change GeoCoordinate struct to accept double rather decimal

+++ SolrExpress.Core/Query/GeoCoordinate.cs
@@ -9,7 +9,7 @@ namespace SolrExpress.Core.Query
         /// </summary>
         /// <param name="latitude">The latitude of the location</param>
         /// <param name="longitude">The longitude of the location</param>
-        public GeoCoordinate(double latitude, double longitude)
+        public GeoCoordinate(decimal latitude, decimal longitude)
         {
             this.Latitude = latitude;
             this.Longitude = longitude;
@@ -18,12 +18,12 @@ namespace SolrExpress.Core.Query
         /// <summary>
         /// The latitude of the location. May range from -90.0 to 90.0.
         /// </summary>
-        public double Latitude;
+        public decimal Latitude;
 
         /// <summary>
         /// The longitude of the location. May range from -180.0 to 180.0.
         /// </summary>
-        public double Longitude;
+        public decimal Longitude;
 
         /// <summary>
         /// Returns a string that represents the current coordinate
